+++ integroBridgeService/src/main/java/com/adobe/campaign/tests/bridge/service/IntegroAPI.java
@@ -9,11 +9,14 @@
 package com.adobe.campaign.tests.bridge.service;
 
 import com.adobe.campaign.tests.bridge.service.exceptions.*;
+import com.adobe.campaign.tests.bridge.service.utils.ServiceTools;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import java.io.File;
+import java.util.HashMap;
+import java.util.Map;
 
 import static spark.Spark.*;
 
@@ -26,11 +29,18 @@ public class IntegroAPI {
     protected static final String ERROR_IBS_CONFIG = "The provided class and method for setting environment variables is not valid.";
     protected static final String ERROR_IBS_RUNTIME = "Problems with payload.";
     public static final String ERROR_CALL_TIMEOUT = "The call you made exceeds the set timeout limit.";
-    public static final String ERRROR_CONTENT_TYPE = "application/problem+json";
+    public static final String ERROR_CONTENT_TYPE = "application/problem+json";
     protected static final String ERROR_AMBIGUOUS_METHOD = "No unique method could be identified that matches your request.";
+    public static final String SYSTEM_UP_MESSAGE = "All systems up";
+    protected enum DeploymentMode {
+            TEST, PRODUCTION
+    }
 
-    public static void
-    startServices(int port) {
+    public static void startServices(int port) {
+
+        if (!ServiceTools.isPortFree(port)) {
+            throw new IBSConfigurationException("The port "+port+" is not currently free.");
+        }
 
         if (Boolean.parseBoolean(ConfigValueHandlerIBS.SSL_ACTIVE.fetchValue())) {
             File l_file = new File(ConfigValueHandlerIBS.SSL_KEYSTORE_PATH.fetchValue());
@@ -43,18 +53,17 @@ public class IntegroAPI {
         }
 
         get("/test", (req, res) -> {
-            res.type("text/plain");
+            res.type("application/json");
+            Map<String, String> status = new HashMap<>();
+            status.put("overALLSystemState",SYSTEM_UP_MESSAGE);
+            status.put("deploymentMode", ConfigValueHandlerIBS.DEPLOYMENT_MODEL.fetchValue());
 
-            StringBuilder sb = new StringBuilder("All systems up "+ ConfigValueHandlerIBS.DEPLOYMENT_MODEL.fetchValue());
-            sb.append("\n");
-            sb.append("Bridge Service Version : ");
-            sb.append(ConfigValueHandlerIBS.PRODUCT_VERSION.fetchValue());
+            status.put("bridgeServiceVersion", ConfigValueHandlerIBS.PRODUCT_VERSION.fetchValue());
             if (ConfigValueHandlerIBS.PRODUCT_USER_VERSION.isSet()) {
-                sb.append("\n");
-                sb.append("Product user version : ");
-                sb.append(ConfigValueHandlerIBS.PRODUCT_USER_VERSION.fetchValue());
+                status.put("hostVersion", ConfigValueHandlerIBS.PRODUCT_USER_VERSION.fetchValue());
             }
-            return sb.toString();
+
+            return BridgeServiceFactory.transformMapTosResult(status);
         });
 
         post("/service-check", (req, res) -> {
@@ -70,84 +79,63 @@ public class IntegroAPI {
 
         after((req, res) -> res.type("application/json"));
 
-        exception( JsonProcessingException.class, (e, req, res) -> {
-
-            StringBuilder response = new StringBuilder();
-            response.append(ERROR_JSON_TRANSFORMATION);
-            response.append("\n");
-            response.append(e.getMessage());
-            res.status(404);
-            res.type(ERRROR_CONTENT_TYPE);
-
-            res.body(response.toString());
+        exception(JsonProcessingException.class, (e, req, res) -> {
+            int statusCode = 404;
+            res.status(statusCode);
+            res.type(ERROR_CONTENT_TYPE);
+            res.body(BridgeServiceFactory.createExceptionPayLoad(
+                    new ErrorObject(e, ERROR_JSON_TRANSFORMATION, statusCode)));
         });
 
-        exception( AmbiguousMethodException.class, (e, req, res) -> {
-            StringBuilder response = new StringBuilder();
-            response.append(ERROR_AMBIGUOUS_METHOD);
-            response.append("\n");
-            response.append(e.getMessage());
-            res.status(404);
-            res.type(ERRROR_CONTENT_TYPE);
-
-            res.body(response.toString());
+        exception(AmbiguousMethodException.class, (e, req, res) -> {
+            int statusCode = 404;
+            res.status(statusCode);
+            res.type(ERROR_CONTENT_TYPE);
+            res.body(BridgeServiceFactory.createExceptionPayLoad(
+                    new ErrorObject(e, ERROR_AMBIGUOUS_METHOD, statusCode)));
         });
 
-        exception( IBSConfigurationException.class, (e, req, res) -> {
-            StringBuilder response = new StringBuilder();
-            response.append(ERROR_IBS_CONFIG);
-            response.append("\n");
-            response.append(e.getMessage());
-            res.status(500);
-            res.type(ERRROR_CONTENT_TYPE);
+        exception(IBSConfigurationException.class, (e, req, res) -> {
+            int statusCode = 500;
+            res.status(statusCode);
+            res.type(ERROR_CONTENT_TYPE);
 
-            res.body(response.toString());
+            res.body(BridgeServiceFactory.createExceptionPayLoad(new ErrorObject(e, ERROR_IBS_CONFIG, statusCode)));
         });
 
-        exception( IBSRunTimeException.class, (e, req, res) -> {
-            StringBuilder response = new StringBuilder();
-            response.append(ERROR_IBS_RUNTIME);
-            response.append("\n");
-            response.append(e.getMessage());
-            res.status(500);
-            res.type(ERRROR_CONTENT_TYPE);
+        exception(IBSRunTimeException.class, (e, req, res) -> {
+            int statusCode = 500;
+            res.status(statusCode);
+            res.type(ERROR_CONTENT_TYPE);
 
-            res.body(response.toString());
+            res.body(BridgeServiceFactory.createExceptionPayLoad(new ErrorObject(e, ERROR_IBS_RUNTIME, statusCode)));
+
         });
 
-        exception( TargetJavaMethodCallException.class, (e, req, res) -> {
-            StringBuilder response = new StringBuilder();
-            response.append(ERROR_CALLING_JAVA_METHOD);
-            response.append("\n");
+        exception(TargetJavaMethodCallException.class, (e, req, res) -> {
+            int statusCode = 500;
+            res.status(statusCode);
+            res.type(ERROR_CONTENT_TYPE);
 
-            res.status(500);
-            res.type(ERRROR_CONTENT_TYPE);
-
-            response.append(e.getMessage()).append("\n");
-            res.body(response.toString());
+            res.body(BridgeServiceFactory.createExceptionPayLoad(
+                    new ErrorObject(e, ERROR_CALLING_JAVA_METHOD, statusCode)));
         });
 
-        exception( NonExistentJavaObjectException.class, (e, req, res) -> {
-            StringBuilder response = new StringBuilder();
-            response.append(ERROR_JAVA_OBJECT_NOT_FOUND);
-            response.append("\n");
-            response.append(e.getMessage());
-            res.status(404);
-            res.type(ERRROR_CONTENT_TYPE);
+        exception(NonExistentJavaObjectException.class, (e, req, res) -> {
+            int statusCode = 404;
+            res.status(statusCode);
+            res.type(ERROR_CONTENT_TYPE);
 
-
-            res.body(response.toString());
+            res.body(BridgeServiceFactory.createExceptionPayLoad(
+                    new ErrorObject(e, ERROR_JAVA_OBJECT_NOT_FOUND, statusCode)));
         });
 
-        exception( IBSTimeOutException.class, (e, req, res) -> {
-            StringBuilder response = new StringBuilder();
-            response.append(ERROR_CALL_TIMEOUT);
-            response.append("\n");
-            response.append(e.getMessage());
-            res.type(ERRROR_CONTENT_TYPE);
+        exception(IBSTimeOutException.class, (e, req, res) -> {
+            int statusCode = 408;
+            res.status(statusCode);
+            res.type(ERROR_CONTENT_TYPE);
 
-            res.status(408);
-            res.body(response.toString());
+            res.body(BridgeServiceFactory.createExceptionPayLoad(new ErrorObject(e, ERROR_CALL_TIMEOUT, statusCode)));
         });
 
     }